backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
PyPDF2
sentence-transformers
tqdm
numpy
requests
python-docx
reportlab
watchdog